Add script to pre-process text files
Text files used for training a keyboard layout need to be proprocessed to
remove names, typos, OCR errors and other artefacts that can distort the
search. Use a word list file as reference and only allow words contained
in it. Punctuations are preserved. Any missing words are output to a new
wordlist that can be used to update the reference if needed.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>