Save discovered layouts to a database
Use a directory as a database, with the file name generated from the
layout itself as a key. Each entry is a python script containing the
layout formated as list, and a multi-line comment string with
performance metrics of the layout. Eventually this can be extended to
make is a functioning script that can evaluate the stored layout against
other texts.

Also stop evaluating all the reference layouts.

The purpose is to run this script in a loop, potentially multiple
instances concurrently, to find many good layout candidates and
eliminate duplicates.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>