Improve heatmap scoring
Calculate the distance between the actual (normalized) heatmap to the
key weights as the distance between two 30-dimensional vectors where
each key weight is one dimension. However, instead of the difference
of elements calculate their ratio and take its logarithm so that ratios
in either direction get the same magnitude. (e.g. log 2 == -log 1/2)
This weighs overuse and underuse of a key equally badly and
proportionally to the weight of the key.

The distance measure using pythagoras is more sensitive to deviations in
individual dimensions (keys) than a simple average of differences over
all keys.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>