Modified heatmap scoring function
Instead of log a/b, use (1 - a/b), which penalizes underusing keys less
than overusing them. In some cases this can lead to unintuitive key
assignments for very rare symbols where a more frequent symbol gets
assigned to a lower weighted key because it is a closer match. But those
are easy to fix manually.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>